organized files and added more dtos
