+++ server/src/Controllers/hotels.controller.ts
@@ -1,6 +1,7 @@
 import { Request, Response, NextFunction } from "express";
 import Hotel from "../Models/Hotel.Model";
-import NewHotelDTO from "../dtos/new-hotel.dto";
+import NewHotelDTO from "../dtos/hotels/new-hotel.dto";
+import { GetHotelDTO } from "../dtos/hotels/get-hotel.dto";
 export class HotelsController {
   private readonly hotels: typeof Hotel;
   private readonly uuid: () => string;
@@ -12,7 +13,18 @@ export class HotelsController {
   getHotels = async (_req: Request, res: Response, next: NextFunction) => {
     try {
       const hotels = await this.hotels.findAll();
-      res.status(200).send({ hotels }).end();
+      if (!hotels.length) {
+        res.send({ hotels: [] }).end();
+        return;
+      }
+      const hotelsDTO: GetHotelDTO[] = hotels.map((hotel) => ({
+        id: hotel?.id,
+        name: hotel?.name,
+        address: hotel?.address,
+        rating: hotel?.rating,
+        img_url: hotel?.img_url,
+      }));
+      res.status(200).send({ hotels: hotelsDTO }).end();
     } catch (err) {
       next(err);
     }
@@ -23,12 +35,11 @@ export class HotelsController {
       const { name, address, rating } = req.body;
       const id = this.uuid();
       const hotel = await this.hotels.create({ id, name, address, rating });
-        const newHotel: NewHotelDTO = {
-                id: hotel?.id,
-                name: hotel?.name,
-                address: hotel?.address,
-
-            }
+      const newHotel: NewHotelDTO = {
+        id: hotel?.id,
+        name: hotel?.name,
+        address: hotel?.address,
+      };
       res.status(201).send({ hotel: newHotel }).end();
     } catch (err) {
       next(err);
@@ -39,7 +50,18 @@ export class HotelsController {
     try {
       const { id: hotelId } = req.params;
       const hotel = await this.hotels.findOne({ where: { id: hotelId } });
-      res.status(200).send({ hotel }).end();
+      if (!hotel) {
+        res.status(404).send({ hotel: null }).end();
+        return;
+      }
+      const hotelDTO: GetHotelDTO = {
+        id: hotel?.id,
+        name: hotel?.name,
+        address: hotel?.address,
+        rating: hotel?.rating,
+        img_url: hotel?.img_url,
+      };
+      res.status(200).send({ hotel: hotelDTO }).end();
     } catch (err) {
       next(err);
     }
