+++ server/src/dtos/hotels/get-hotel.dto.ts
@@ -1,0 +1,7 @@
+export class GetHotelDTO{
+     id: string;
+     name: string;
+     rating: number;
+     address: string;
+     img_url: string;
+}
