+++ server/src/dtos/hotels/new-hotel.dto.ts
